refactor: move DocumentFormattingEditProvider, DocumentRangeFormattingEditProvider to new file
