+++ src/language/common/lspLanguageFeatures.ts
@@ -9,7 +9,6 @@ import {
 	editor,
 	Uri,
 	Position,
-	Range,
 	CancellationToken
 } from '../../fillers/monaco-editor-core';
 import { fromPosition, toRange, toTextEdit, fromRange } from './CompletionAdapter';
@@ -36,77 +35,7 @@ export * from './DocumentSymbolAdapter';
 
 export * from './DocumentLinkAdapter';
 
-//#region DocumentFormattingEditProvider, DocumentRangeFormattingEditProvider
-
-export interface ILanguageWorkerWithFormat {
-	format(
-		uri: string,
-		range: lsTypes.Range | null,
-		options: lsTypes.FormattingOptions
-	): Promise<lsTypes.TextEdit[]>;
-}
-
-export class DocumentFormattingEditProvider<T extends ILanguageWorkerWithFormat>
-	implements languages.DocumentFormattingEditProvider
-{
-	constructor(private _worker: WorkerAccessor<T>) {}
-
-	public provideDocumentFormattingEdits(
-		model: editor.IReadOnlyModel,
-		options: languages.FormattingOptions,
-		token: CancellationToken
-	): Promise<languages.TextEdit[] | undefined> {
-		const resource = model.uri;
-
-		return this._worker(resource).then((worker) => {
-			return worker
-				.format(resource.toString(), null, fromFormattingOptions(options))
-				.then((edits) => {
-					if (!edits || edits.length === 0) {
-						return;
-					}
-					return edits.map<languages.TextEdit>(toTextEdit);
-				});
-		});
-	}
-}
-
-export class DocumentRangeFormattingEditProvider<T extends ILanguageWorkerWithFormat>
-	implements languages.DocumentRangeFormattingEditProvider
-{
-	readonly canFormatMultipleRanges = false;
-
-	constructor(private _worker: WorkerAccessor<T>) {}
-
-	public provideDocumentRangeFormattingEdits(
-		model: editor.IReadOnlyModel,
-		range: Range,
-		options: languages.FormattingOptions,
-		token: CancellationToken
-	): Promise<languages.TextEdit[] | undefined> {
-		const resource = model.uri;
-
-		return this._worker(resource).then((worker) => {
-			return worker
-				.format(resource.toString(), fromRange(range), fromFormattingOptions(options))
-				.then((edits) => {
-					if (!edits || edits.length === 0) {
-						return;
-					}
-					return edits.map<languages.TextEdit>(toTextEdit);
-				});
-		});
-	}
-}
-
-function fromFormattingOptions(options: languages.FormattingOptions): lsTypes.FormattingOptions {
-	return {
-		tabSize: options.tabSize,
-		insertSpaces: options.insertSpaces
-	};
-}
-
-//#endregion
+export * from './DocumentFormattingEditProvider';
 
 //#region DocumentColorAdapter
